+++ src/containers/App/App.test.js
@@ -1,16 +1,15 @@
-import { cleanup, fireEvent, render, waitFor } from '@testing-library/react';
+import '@testing-library/jest-dom/extend-expect';
+import { fireEvent, render } from '@testing-library/react';
 import FeatureProvider, { setFeatures } from 'feature-provider';
 import React from 'react';
 import Footer from '../../components/Footer/Footer';
-import NavBar from '../../components/NavBar/NavBar';
+import { DANGER, SUCCESS } from '../../constants';
 import features from '../../features';
 import { useAuth0 } from '../../react-auth0-wrapper';
-import makeInitialState from '../../store/makeInitialState';
-import makeReducer from '../../store/makeReducer';
+import makeReducer, { types } from '../../store/makeReducer';
 import App from './App';
 
 jest.mock('../../components/Footer/Footer');
-jest.mock('../../components/NavBar/NavBar');
 jest.mock('../../react-auth0-wrapper');
 
 const feature = setFeatures(features);
@@ -22,16 +21,9 @@ const user = {
   sub: 'twitter|1183836409850814464'
 };
 
-describe.skip('App', () => {
+describe('<App />', () => {
   beforeEach(() => {
-    fetch.mockResponseOnce(JSON.stringify({ data: { message: 'baz' } }));
     Footer.mockImplementation(() => <div>Footer</div>);
-    NavBar.mockImplementation(() => <div>NavBar</div>);
-  });
-
-  afterEach(() => {
-    fetch.resetMocks();
-    cleanup();
   });
 
   it('loading', () => {
@@ -39,136 +31,105 @@ describe.skip('App', () => {
       loading: true
     }));
 
-    const { container } = render(
+    const initialState = {};
+
+    const { getByText } = render(
       <FeatureProvider features={features}>
-        <App
-          initialState={makeInitialState({ feature })}
-          reducer={makeReducer(feature)}
-        />
+        <App initialState={initialState} reducer={makeReducer(feature)} />
       </FeatureProvider>
     );
 
-    expect(container).toMatchSnapshot();
+    expect(getByText('...')).toBeInTheDocument();
   });
 
-  it('unauthenticated user', () => {
+  it('not authenticated', () => {
     useAuth0.mockImplementation(() => ({
       isAuthenticated: false
     }));
 
-    const { container } = render(
-      <FeatureProvider features={features}>
-        <App
-          initialState={makeInitialState({ feature })}
-          reducer={makeReducer(feature)}
-        />
-      </FeatureProvider>
-    );
-
-    expect(container).toMatchSnapshot();
-  });
-
-  it('log in', () => {
-    const mockLoginWithRedirect = jest.fn();
-    useAuth0.mockImplementation(() => ({
-      loginWithRedirect: mockLoginWithRedirect
-    }));
+    const initialState = {};
 
     const { getByTestId } = render(
       <FeatureProvider features={features}>
-        <App
-          initialState={makeInitialState({ feature })}
-          reducer={makeReducer(feature)}
-        />
+        <App initialState={initialState} reducer={makeReducer(feature)} />
       </FeatureProvider>
     );
 
-    fireEvent.click(getByTestId('login'));
-
-    expect(mockLoginWithRedirect).toBeCalled();
+    expect(getByTestId('login')).toBeInTheDocument();
   });
 
-  it('generate tweetstorm and dismiss notification', async () => {
-    fetch.mockResponseOnce(JSON.stringify({ data: { message: 'qux' } })); // INFO: adding a second response
+  it('authenticated', () => {
     useAuth0.mockImplementation(() => ({
       isAuthenticated: true,
       user
     }));
 
-    const { container, getByTestId } = render(
+    const initialState = {};
+
+    const { getByTestId } = render(
       <FeatureProvider features={features}>
-        <App
-          initialState={makeInitialState({ feature })}
-          reducer={makeReducer(feature)}
-        />
+        <App initialState={initialState} reducer={makeReducer(feature)} />
       </FeatureProvider>
     );
-    fireEvent.change(getByTestId('source'), { target: { value: 'foo' } });
-    await waitFor(() => {
-      fireEvent.click(getByTestId('tweet'));
-    });
-    await waitFor(() => {
-      fireEvent.click(getByTestId('dismiss'));
-    });
 
-    expect(container).toMatchSnapshot();
+    expect(getByTestId('logout')).toBeInTheDocument();
   });
-});
 
-describe.skip('App - errors', () => {
-  it('fetchHealth error', () => {
-    fetch.resetMocks();
-    fetch.mockRejectOnce(new Error('Oh Noes!'));
-    Footer.mockImplementation(() => <div>Footer</div>);
-    NavBar.mockImplementation(() => <div>NavBar</div>);
+  it('notification danger', () => {
+    useAuth0.mockImplementation(() => ({
+      isAuthenticated: true,
+      user
+    }));
+
+    const initialState = {
+      notification: { message: 'Oh Noes!', type: DANGER }
+    };
+
+    const mockReducer = jest.fn((state, action) =>
+      makeReducer(feature)(state, action)
+    );
+
+    const { getByTestId } = render(
+      <FeatureProvider features={features}>
+        <App initialState={initialState} reducer={mockReducer} />
+      </FeatureProvider>
+    );
+
+    fireEvent.click(getByTestId('dismiss'));
+
+    expect(mockReducer).toBeCalledTimes(2);
+    expect(mockReducer.mock.calls[1][1]).toEqual({ type: types.DISMISS_TOAST }); // [1][1] === [the send time][the action parameter]
+  });
+
+  it('notification success', () => {
+    jest.useFakeTimers();
 
     useAuth0.mockImplementation(() => ({
-      isAuthenticated: false
+      isAuthenticated: true,
+      user
     }));
+
+    const initialState = {
+      notification: {
+        link:
+          'https://twitter.com/1183836409850814464/status/1230233654959407104',
+        type: SUCCESS
+      }
+    };
+
     const mockReducer = jest.fn((state, action) =>
       makeReducer(feature)(state, action)
     );
 
     render(
       <FeatureProvider features={features}>
-        <App
-          initialState={makeInitialState({ feature })}
-          reducer={mockReducer}
-        />
+        <App initialState={initialState} reducer={mockReducer} />
       </FeatureProvider>
     );
 
-    // TODO: add assertion
-  });
+    jest.advanceTimersByTime(4000);
 
-  it('fetchTweetstorm error', async () => {
-    fetch.resetMocks();
-    fetch.mockResponseOnce(JSON.stringify({ data: { message: 'baz' } }));
-    fetch.mockRejectOnce(new Error('Oh Noes!'));
-    Footer.mockImplementation(() => <div>Footer</div>);
-    NavBar.mockImplementation(() => <div>NavBar</div>);
-
-    useAuth0.mockImplementation(() => ({
-      isAuthenticated: true,
-      user
-    }));
-    const mockReducer = jest.fn((state, action) =>
-      makeReducer(feature)(state, action)
-    );
-
-    const { getByTestId } = render(
-      <FeatureProvider features={features}>
-        <App
-          initialState={makeInitialState({ feature })}
-          reducer={mockReducer}
-        />
-      </FeatureProvider>
-    );
-    fireEvent.change(getByTestId('source'), { target: { value: 'foo' } });
-    await waitFor(() => {
-      fireEvent.click(getByTestId('tweet'));
-    });
-
-    // TODO: add assertion
+    expect(mockReducer).toBeCalledTimes(2);
+    expect(mockReducer.mock.calls[1][1]).toEqual({ type: types.DISMISS_TOAST }); // [1][1] === [the send time][the action parameter]
   });
 });
